requirements.txt
promptflow
promptflow-tools
azure-cosmos
azure-ai-ml
azure-ai-resources
azure-search-documents==11.4.0
openai
requests
python-dotenv
